refactor: move bulk transforms into transforms/merge.py
The dataset-level (bulk) transforms physically lived in the legacy
pipeline/transform.py, and transforms/merge.py was a thin re-export over
them. Invert that: the implementations now live in transforms/merge.py
(their natural home in the new Reader/Writer/Transform structure), and
pipeline/transform.py is a backward-compatibility shim that re-exports
from the new location so existing imports keep working while the
pipeline package is retired.

- transforms/merge.py: owns write_staging, merge_staging_files,
  merge_comments_partitioned, update_comments_index, partition_comments,
  build_feed_summary, build_agency_rollups.
- pipeline/transform.py: re-export shim (deprecated).
- Repoint first-party importers (sources/parquet.py, tests) at the new
  home; fix the transforms/base.py doc pointer.

A first step toward deprecating the pipeline/ package; the remaining
modules (extract, load, download_r2, upload_r2, the Prefect flow) are
left for a follow-up.
